backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "civicpulse-api"
version = "0.1.0"
description = "FastAPI backend for Civic Pulse"
authors = [{ name = "Civic Pulse" }]
dependencies = [
    "fastapi>=0.112",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "azure-identity>=1.17",
    "azure-ai-openai>=1.0.0b7",
    "azure-search-documents>=11.5",
    "semantic-kernel>=1.13",
    "azure-ai-contentsafety>=1.0.0b1",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.uvicorn]
app = "civicpulse_api.main:app"
host = "0.0.0.0"
port = 8000
reload = true
